requirements.txt
google-api-python-client==2.120.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
appium-python-client==3.1.0
selenium==4.20.0
